+++ service/controller/app/v1/resource/appnamespace/create.go
@@ -1,4 +1,4 @@
-package namespace
+package appnamespace
 
 import "context"
