+++ service/controller/app/v1/resource/appnamespace/delete.go
@@ -1,4 +1,4 @@
-package namespace
+package appnamespace
 
 import (
 	"context"
